Restructuration and first modifications of the project
Added new files and folders for the global structure of the project
Added in error.c some error exit
Author: Mat1eub
